requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.4.0
djangorestframework==3.15.2
pytz==2019.3
sqlparse==0.4.4
tzdata==2023.3